fix warning in block_cache.cpp

+++ src/block_cache.cpp
@@ -617,7 +617,7 @@ cached_piece_entry* block_cache::allocate_piece(disk_io_job const* j, std::uint1
 		pe.piece = j->piece;
 		pe.storage = j->storage;
 		pe.expire = aux::time_now();
-		pe.blocks_in_piece = aux::numeric_cast<std::uint64_t>(blocks_in_piece);
+		pe.blocks_in_piece = aux::numeric_cast<std::uint16_t>(blocks_in_piece);
 
 		pe.blocks.reset(new (std::nothrow) cached_block_entry[std::size_t(blocks_in_piece)]);
 		if (!pe.blocks) return nullptr;
